macro:fix rename lifesensor_macro to lifesensor_common
