+++ components/lifesensor_common/CMakeLists.txt
