+++ components/lifesensor_common/include/macro/queue.h
