+++ components/lifesensor_common/include/macro/task.h
