+++ components/ulp_adc/src/ulp.c
@@ -4,7 +4,7 @@
 
 #include "spo2.h"
 #include "ulp.h"
-#include "macro_queue.h"
+#include "macro/queue.h"
 
 
 /*
